Fixed a bug in neighbour tile bbox transformation
Since the tiles weren't being transformed to the layer sref, they were
lining up less and less the closer you zoomed and therefore would not
find the features in the boxes. This is the actual root cause of the
previously "fixed" issue in Iceland, which has now been fixed properly.
(Iceland was the only place it showed up at the world-level zoom.)